PROJECTONE/requirements.txt
aiohttp==3.7.4.post0
alembic==1.6.5
anyio==3.2.1
apispec==3.3.2
argcomplete==1.12.3
asgiref==3.4.1
asn1crypto==1.4.0
asttokens==2.0.4
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
base58==2.1.0
bitarray==1.2.2
black==21.9b0
blinker==1.4
boto3==1.17.49
botocore==1.20.49
cattrs==1.5.0
certifi==2021.10.8
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.7
cytoolz==0.11.0
dataclassy==0.11.0
defusedxml==0.7.1
dill==0.3.4
dj-database-url==0.5.0
Django==3.2.8
django-bower==5.2.0
django-crispy-forms==1.13.0
django-ipfs-storage==0.0.4
django-model-utils==4.2.0
django-notifications-hq==1.6.0
django-scheduler==0.9.3
django-storages==1.11.1
dnspython==2.1.0
docutils==0.16
eip712==0.1.0
email-validator==1.1.3
eth-abi==2.1.1
eth-account==0.5.6
eth-brownie==1.17.0
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
execnet==1.9.0
Flask==1.1.4
Flask-AppBuilder==3.3.1
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.16
gunicorn==20.1.0
h11==0.12.0
hexbytes==0.2.2
httpcore==0.13.6
httpx==0.18.2
hypothesis==6.23.2
icalendar==4.0.7
idna==3.2
importlib-resources==1.5.0
inflection==0.5.0
iniconfig==1.1.1
ipfsapi==0.4.4
ipfshttpclient==0.8.0a2
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
lru-dict==1.1.7
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
numpy==1.21.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.0
pandas==1.3.0
parsimonious==0.8.1
pathspec==0.9.0
pendulum==2.1.2
Pillow==8.2.0
pipx==0.16.4
platformdirs==2.4.0
pluggy==1.0.0
prison==0.1.3
prompt-toolkit==3.0.20
protobuf==3.18.1
psutil==5.8.0
psycopg2==2.8.6
py==1.10.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pycparser==2.20
pycryptodome==3.11.0
Pygments==2.10.0
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
pyparsing==2.4.7
pypiwin32==223
pyreadline3==3.3
pyrsistent==0.18.0
pysha3==1.0.2
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-daemon==2.3.0
python-dateutil==2.8.1
python-dotenv==0.16.0
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pythx==1.6.1
pytz==2021.1
pytzdata==2020.1
pywin32==300
PyYAML==5.4.1
regex==2021.10.8
repoze.lru==0.7
requests==2.26.0
rfc3986==1.5.0
rich==10.4.0
rlp==1.2.0
s3transfer==0.3.6
scrypt==0.8.18
semantic-version==2.8.5
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
sqlparse==0.4.1
swagger-ui-bundle==0.0.8
swapper==1.1.2.post1
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
toolz==0.11.1
tqdm==4.62.3
trie==1.4.0
typing-extensions==3.10.0.2
unicodecsv==0.14.1
urllib3==1.26.7
userpath==1.7.0
varint==1.0.2
vvm==0.1.0
vyper==0.3.0
wcwidth==0.2.5
web3==5.24.0
websockets==9.1
Werkzeug==1.0.1
wrapt==1.13.1
WTForms==2.3.3
yarl==1.7.0
zope.interface==5.4.0
